requirements.txt
discord.py>=2.3,<3.0
faster-whisper>=1.0,<2.0
anthropic>=0.40,<1.0
aiohttp>=3.9,<4.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
google-api-python-client>=2.0,<3.0
google-auth>=2.0,<3.0
